refactor(bsw): consolidate BswBehavior leaf package into BswBehavior.py (Rule 0007)
Rule 0007: a leaf package (no subpackages) defines its classes in Pkg.py, and
never in a class-named submodule. BswBehavior/ had no subpackages, so its 39
classes moved from BswBehavior/__init__.py into BswModuleTemplate/BswBehavior.py;
BswBehavior/BswInterruptEvent.py (a class-named submodule) is folded in and the
directory removed.

Rule 0007 shadowing note: creating BswBehavior.py while BswBehavior/ existed would
have made every class in BswBehavior/*.py dead code, so both changes land together.
Imports of ...BswBehavior are unchanged (valid for package or module); the three
references to the class-named submodule are retargeted (arxml_writer, unit test,
deferred import inside createBswInterruptEvent). BswInterruptEvent leaves
KNOWN_NAME_COLLISION_CLASSES, and the stale models/__init__.py note is corrected.

Verified: 7159 unit tests pass, flake8+ruff clean, black-check clean, integration
round-trip PASS.
